# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface AssessRequest {
  msme_id: string;
}

export interface Pillar {
  name: string;
  score: number;
  reasons: string[];
}

export interface Confidence {
  level: string;
  raise_by: string;
}

export interface Repayment {
  sustainable_emi: number;
  basis: string;
}

export interface Flags {
  consistency_alert: boolean;
  detail: string;
}

export interface Forecast {
  months: string[];
  projected_net_surplus: number[];
}

export interface CardResponse {
  msme_id: string;
  overall_score: number;
  rating_band: string;
  pillars: Pillar[];
  confidence: Confidence;
  repayment: Repayment;
  flags: Flags;
  forecast: Forecast;
}

export interface ScoreBucket {
  bucket: string;
  count: number;
}

export interface SectorConcentration {
  sector: string;
  count: number;
  avgScore: number;
}

export interface PortfolioResponse {
  scoreDistribution: ScoreBucket[];
  sectorConcentration: SectorConcentration[];
  expectedDefaultEstimate: number;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  lender: 'lender',
  borrower: 'borrower',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  msme_id?: string;
}

export interface AuthResponse {
  user: User;
}

export type SignupRequestRole = typeof SignupRequestRole[keyof typeof SignupRequestRole];


export const SignupRequestRole = {
  lender: 'lender',
  borrower: 'borrower',
} as const;

export interface SignupRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  name: string;
  role: SignupRequestRole;
  msme_id?: string;
}

export interface LoginRequest {
  email: string;
  /** @minLength 1 */
  password: string;
}