# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scoring
    description: MSME assessment, scorecard, and portfolio operations
  - name: auth
    description: Authentication operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /assess:
    post:
      operationId: assess
      tags: [scoring]
      summary: Assess an MSME
      description: Compute the full scorecard for one MSME and cache the result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessRequest"
      responses:
        "200":
          description: The full scorecard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user is not a lender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unknown MSME
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /card/{msme_id}:
    get:
      operationId: getCard
      tags: [scoring]
      summary: Get a cached scorecard
      description: Return the most recently assessed scorecard for an MSME.
      parameters:
        - name: msme_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The cached scorecard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user is not a lender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No cached scorecard for this MSME
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /portfolio:
    get:
      operationId: getPortfolio
      tags: [scoring]
      summary: Portfolio analytics
      description: Aggregate score distribution, sector concentration, and expected default across all MSMEs.
      responses:
        "200":
          description: Portfolio analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user is not a lender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/scorecard:
    get:
      operationId: getMyScorecard
      tags: [scoring]
      summary: Get my own scorecard
      description: Returns the scorecard for the MSME linked to the authenticated borrower.
      responses:
        "200":
          description: The scorecard for the authenticated borrower's linked MSME
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated user is not a borrower
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No MSME linked to this account, or the linked MSME is unknown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      description: Creates a new user account and starts a session (sets httpOnly auth cookies).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: An account with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid signup payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      description: Verifies credentials and starts a session (sets httpOnly auth cookies).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refresh
      tags: [auth]
      summary: Refresh the session
      description: Exchanges a valid refresh-token cookie for a new access/refresh token pair.
      responses:
        "200":
          description: Session refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Missing, invalid, expired, or revoked refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      description: Revokes the current refresh token and clears auth cookies.
      responses:
        "204":
          description: Logged out
  /auth/me:
    get:
      operationId: me
      tags: [auth]
      summary: Get the current user
      description: Returns the authenticated user, based on the access-token cookie.
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    AssessRequest:
      type: object
      properties:
        msme_id:
          type: string
      required:
        - msme_id
    Pillar:
      type: object
      properties:
        name:
          type: string
        score:
          type: number
        reasons:
          type: array
          items:
            type: string
      required:
        - name
        - score
        - reasons
    Confidence:
      type: object
      properties:
        level:
          type: string
        raise_by:
          type: string
      required:
        - level
        - raise_by
    Repayment:
      type: object
      properties:
        sustainable_emi:
          type: number
        basis:
          type: string
      required:
        - sustainable_emi
        - basis
    Flags:
      type: object
      properties:
        consistency_alert:
          type: boolean
        detail:
          type: string
      required:
        - consistency_alert
        - detail
    Forecast:
      type: object
      properties:
        months:
          type: array
          items:
            type: string
        projected_net_surplus:
          type: array
          items:
            type: number
      required:
        - months
        - projected_net_surplus
    CardResponse:
      type: object
      properties:
        msme_id:
          type: string
        overall_score:
          type: number
        rating_band:
          type: string
        pillars:
          type: array
          items:
            $ref: "#/components/schemas/Pillar"
        confidence:
          $ref: "#/components/schemas/Confidence"
        repayment:
          $ref: "#/components/schemas/Repayment"
        flags:
          $ref: "#/components/schemas/Flags"
        forecast:
          $ref: "#/components/schemas/Forecast"
      required:
        - msme_id
        - overall_score
        - rating_band
        - pillars
        - confidence
        - repayment
        - flags
        - forecast
    ScoreBucket:
      type: object
      properties:
        bucket:
          type: string
        count:
          type: integer
      required:
        - bucket
        - count
    SectorConcentration:
      type: object
      properties:
        sector:
          type: string
        count:
          type: integer
        avgScore:
          type: number
      required:
        - sector
        - count
        - avgScore
    PortfolioResponse:
      type: object
      properties:
        scoreDistribution:
          type: array
          items:
            $ref: "#/components/schemas/ScoreBucket"
        sectorConcentration:
          type: array
          items:
            $ref: "#/components/schemas/SectorConcentration"
        expectedDefaultEstimate:
          type: number
      required:
        - scoreDistribution
        - sectorConcentration
        - expectedDefaultEstimate
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [lender, borrower]
        msme_id:
          type: string
      required:
        - id
        - email
        - name
        - role
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user
    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
        role:
          type: string
          enum: [lender, borrower]
        msme_id:
          type: string
      required:
        - email
        - password
        - name
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required:
        - email
        - password