# lib/api-zod/src/generated/types/signupRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SignupRequestRole } from './signupRequestRole';

export interface SignupRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  name: string;
  role: SignupRequestRole;
  msme_id?: string;
}