# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  msme_id?: string;
}